+++ src/extension-firefox/storage.js
@@ -38,8 +38,10 @@ const defaultShortcutOptions = {
 };
 
 const defaultMenuOptions = {
-  headings: 'main-only h1 h2',
+  headings: 'h1 h2',
   landmarks: 'main search navigation complementary',
+  excludeSmallHeadings: true,
+  excludeLandmarksWithoutNames: true
 };
 
 const defaultHighlightOptions = {