+++ src/skipto.js
@@ -18,7 +18,7 @@
  * page or a bookmarlet will be ignored.
  */
 
-import SkipToContent576 from './skiptoContent.js';
+import SkipToContent580 from './skiptoContent.js';
 import DebugLogging  from './debug.js';
 
 import {
@@ -130,7 +130,7 @@ debug.flag = false;
         if (!isExtensionLoaded) {
           if (!isBookmarkletLoaded) {
             removePageSkipTo();
-            window.customElements.define(BOOKMARKLET_ELEMENT_NAME, SkipToContent576);
+            window.customElements.define(BOOKMARKLET_ELEMENT_NAME, SkipToContent580);
             skipToContentElem = document.createElement(BOOKMARKLET_ELEMENT_NAME);
             skipToContentElem.setAttribute('version', skipToContentElem.version);
             skipToContentElem.setAttribute('type', type);
@@ -146,7 +146,7 @@ debug.flag = false;
         if (!isExtensionLoaded) {
           removePageSkipTo();
           removeBookmarkletSkipTo();
-          window.customElements.define(EXTENSION_ELEMENT_NAME, SkipToContent576);
+          window.customElements.define(EXTENSION_ELEMENT_NAME, SkipToContent580);
           skipToContentElem = document.createElement(EXTENSION_ELEMENT_NAME);
           skipToContentElem.setAttribute('version', skipToContentElem.version);
           skipToContentElem.setAttribute('type', type);
@@ -159,7 +159,7 @@ debug.flag = false;
 
       default:
         if (!isPageLoaded && !isBookmarkletLoaded && !isExtensionLoaded) {
-          window.customElements.define(PAGE_SCRIPT_ELEMENT_NAME, SkipToContent576);
+          window.customElements.define(PAGE_SCRIPT_ELEMENT_NAME, SkipToContent580);
           skipToContentElem = document.createElement(PAGE_SCRIPT_ELEMENT_NAME);
           skipToContentElem.setAttribute('version', skipToContentElem.version);
           skipToContentElem.setAttribute('type', type);