+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Version history for SkipTo.js
 
-## Version 5.7.6
+## Version 5.8.0
 * Improves labels for group labels for headings and landmarks
 * Added 'landmarkOneGroupLabel' and 'headingOneGroupLabel' properties to configuration
 * Changed default headers to be any h1 or h2 on the page
+* Exclude SkipTo.js components from analysis
+* New option to exclude landmarks without accessible names if there are more than 3 landmarks of that type (default is to exclude), this mimics screen reader behaviors
+* New option to exclude headings that are off screen or hidden (default is to exclude), this mimics screen reader behaviors
+* Update accessible name calculation to exclude hidden :before and :after content
 
 ## Version 5.7.5
 * Fixing bug in banner and contentinfo navigation in menu